requirements.txt
Django==4.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
Pillow>= 9.3
python-dotenv==0.19.2
pytz==2019.3
requests==2.28.2